requirements.txt
fastapi
uvicorn
pydantic
pillow
slowapi
python-dotenv